Youtube/Youtube-Backend/requirements.txt
# requirements.txt
flask>=2.3.0
yt-dlp>=2023.10.13
flask-socketio>=5.3.0
python-socketio>=5.9.0
eventlet>=0.33.0
mutagen>=1.46.0
Pillow>=10.0.0
requests>=2.31.0
werkzeug>=2.3.0